Use zlib instead of gzip
The reason to make this change is that I had problems with some URL's and I saw that the problem whas the Gunzip.

I saw that the "gzip" http content is the same that we can found in this stackoverflow post: http://stackoverflow.com/a/3947241/1344260

And it fails too, here you  have my test/problem: http://pastebin.com/3rvmCEkU

It do a IOError: CRC check failed 0xac557afc != 0x2338b236L
and torrenter fails.

After that, I read this message and test the solution: http://stackoverflow.com/a/13692010/1344260

And its working. So I changed the code.